refactor(checker): split line-cap helper shards
